requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
torch
llama-index[llms]>=0.10.28
llama-index-llms-ollama
llama-index-embeddings-huggingface
sacremoses 
protobuf
requests
evaluate
